+++ scripts/scrape_ollama_library.py
@@ -165,7 +165,7 @@ def save_data_to_sqlite(models_info):
             id INTEGER PRIMARY KEY AUTOINCREMENT,
             model_id INTEGER,
             num_params TEXT,
-            size REAL,
+            size INTEGER,
             FOREIGN KEY(model_id) REFERENCES models(id)
         )
     ''')
